back-end/requirements.txt
Django
torch==2.3.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pgvector
psycopg2-binary
gunicorn
requests
python-dotenv
sentence-transformers
PyMuPDF
PyPDF2
Pillow
celery
redis
supabase
dj_database_url
weasyprint
reportlab
qrcode